+++ backend/src/main/kotlin/xyz/funkybit/apps/api/FaucetRoutes.kt
@@ -18,7 +18,6 @@ import xyz.funkybit.apps.api.model.ReasonCode
 import xyz.funkybit.apps.api.model.RequestProcessingError
 import xyz.funkybit.apps.api.model.errorResponse
 import xyz.funkybit.core.blockchain.ChainManager
-import xyz.funkybit.core.model.BitcoinAddress
 import xyz.funkybit.core.model.EvmAddress
 import xyz.funkybit.core.model.Symbol
 import xyz.funkybit.core.model.TxHash
@@ -77,26 +76,8 @@ class FaucetRoutes(private val faucetMode: FaucetMode) {
                         )
 
                         else -> {
-                            val amount = BigDecimal("1")
-
-                            logger.debug { "Sending $amount ${symbol.name} to ${payload.address.value}" }
-
-                            val amountInFundamentalUnits = amount.movePointRight(symbol.decimals.toInt()).toBigInteger()
-                            val txHash = if (symbol.faucetSupported(faucetMode)) {
-                                when (val tokenContractAddress = symbol.contractAddress) {
-                                    null -> blockchainClient.sendNativeDepositTx(
-                                        payload.address,
-                                        amountInFundamentalUnits,
-                                    )
-
-                                    is EvmAddress -> blockchainClient.sendMintERC20Tx(
-                                        tokenContractAddress,
-                                        payload.address,
-                                        amountInFundamentalUnits,
-                                    )
-
-                                    is BitcoinAddress -> TODO()
-                                }
+                            val (amountInFundamentalUnits, txHash) = if (symbol.faucetSupported(faucetMode)) {
+                                blockchainClient.faucetTransfer(symbol, payload.address)
                             } else {
                                 throw RequestProcessingError(
                                     Status.UNPROCESSABLE_ENTITY,
@@ -110,7 +91,7 @@ class FaucetRoutes(private val faucetMode: FaucetMode) {
 
                             Response(Status.OK).with(
                                 responseBody of FaucetApiResponse(
-                                    chainId = blockchainClient.chainId,
+                                    chainId = symbol.chainId.value,
                                     txHash = txHash,
                                     symbol = payload.symbol,
                                     amount = amountInFundamentalUnits,
